+++ lib/paper_trail.rb
@@ -2,7 +2,9 @@ require 'paper_trail/config'
 require 'paper_trail/controller'
 require 'paper_trail/has_paper_trail'
 require 'paper_trail/version'
+
 require 'paper_trail/serializers/yaml'
+require 'paper_trail/serializers/json'
 
 # PaperTrail's module methods can be called in both models and controllers.
 module PaperTrail